+++ Engine/Code/System/MemoryTools/MemoryHelper.h
@@ -5,7 +5,7 @@
 /// 联系作者：94458936@qq.com
 ///
 /// 标准：std:c++20
-/// 版本：1.0.0.3 (2023/12/21 17:43)
+/// 版本：1.0.0.7 (2024/03/04 15:58)
 
 #ifndef SYSTEM_MEMORY_TOOLS_MEMORY_HELPER_H
 #define SYSTEM_MEMORY_TOOLS_MEMORY_HELPER_H
@@ -15,11 +15,11 @@
 #include "Fwd/MemoryToolsFlagsFwd.h"
 #include "System/Windows/Using/WindowsUsing.h"
 
-// 内存工具函数。
+/// 内存工具函数。
 namespace System
 {
     void SYSTEM_DEFAULT_DECLARE MemorySet(void* bytes, int value, int byteSize) noexcept;
-    void SYSTEM_DEFAULT_DECLARE MemoryCopy(void* dst, const void* src, uint32_t count) noexcept;
+    void SYSTEM_DEFAULT_DECLARE MemoryCopy(void* dst, const void* src, size_t count) noexcept;
     MAYBE_NULLPTR SYSTEM_DEFAULT_DECLARE void* MallocMemory(size_t size) noexcept;
     void SYSTEM_DEFAULT_DECLARE FreeMemory(void* memory) noexcept;
